Add generators for Events blocks
