+++ generators/generators-scene.js
@@ -1,5 +1,4 @@
 import * as Blockly from "blockly";
-import { javascriptGenerator } from "blockly/javascript";
 import { meshMap, meshBlockIdMap, generateUniqueId } from "./mesh-state.js";
 import {
   getFieldValue,
